Allow for parametric types.
The type identifier allocation is massive. Reserve five bits to allow for 32
fully distinct types, and allow the rest of the field to be used for parametric
information. Two potential usages are calling convention information for code
pointers and size information for types potentially allocated in blocks of more
than one workspace unit.